requirements.txt
torch>=2.0.0
torchvision
transformers
datasets
evaluate
scikit-learn
matplotlib
pillow